+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Include chart name when installing with a custom release name ([#227], [#228]).
+
+[#227]: https://github.com/stackabletech/superset-operator/pull/227
+[#228]: https://github.com/stackabletech/superset-operator/pull/228
+
 ## [0.5.0] - 2022-06-30
 
 ### Added