+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Include chart name when installing with a custom release name ([#227], [#228]).
 - Orphaned resources are deleted ([#255]).
+- `operator-rs` `0.22.0` -> `0.25.0` ([#255]).
 
 [#227]: https://github.com/stackabletech/superset-operator/pull/227
 [#228]: https://github.com/stackabletech/superset-operator/pull/228